Вынес работу с вебвью в отдельный модуль.
